+++ orm/index.go
@@ -2,7 +2,8 @@ package orm
 
 import (
 	"bytes"
-	math "math"
+	"encoding/hex"
+	"math"
 
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/errors"
@@ -294,6 +295,8 @@ func (i compactIndex) Query(db weave.ReadOnlyKVStore, mod string, data []byte) (
 			return nil, err
 		}
 		return i.loadRefs(db, refs)
+	case weave.RangeQueryMod:
+		return nil, errors.Wrap(errors.ErrHuman, "not implemented: "+mod)
 	default:
 		return nil, errors.Wrap(errors.ErrHuman, "not implemented: "+mod)
 	}
@@ -475,11 +478,11 @@ const nativeIdxPrefix = "_x."
 // NewNativeIndex returns an index implementation that is using a database
 // native storage and query in order to maintain and provide access to an
 // index.
-func NewNativeIndex(name string, indexer MultiKeyIndexer, refKey func([]byte) []byte) Index {
+func NewNativeIndex(name string, indexer MultiKeyIndexer, dbKey func([]byte) []byte) Index {
 	return &nativeIndex{
 		name:    name,
 		indexer: indexer,
-		refKey:  refKey,
+		dbKey:   dbKey,
 	}
 }
 
@@ -488,9 +491,9 @@ func NewNativeIndex(name string, indexer MultiKeyIndexer, refKey func([]byte) []
 type nativeIndex struct {
 	name    string
 	indexer MultiKeyIndexer
-	// refKey is a function that for given entity ID returns that entity
+	// dbKey is a function that for given entity ID returns that entity
 	// database key.
-	refKey func([]byte) []byte
+	dbKey func([]byte) []byte
 }
 
 func (ix *nativeIndex) Name() string {
@@ -584,7 +587,12 @@ func (ix *nativeIndex) Keys(db weave.ReadOnlyKVStore, value []byte) weave.Iterat
 		return &failedIterator{err: err}
 	}
 
-	return &nativeIndexIterator{dbit: it}
+	return &nativeIndexIterator{
+		dbit: it,
+		// Keys method must return keys not prefixed by the bucket
+		// name.
+		dbKey: func(b []byte) []byte { return b },
+	}
 }
 
 func (ix *nativeIndex) Query(db weave.ReadOnlyKVStore, mod string, data []byte) ([]weave.Model, error) {
@@ -596,7 +604,7 @@ func (ix *nativeIndex) Query(db weave.ReadOnlyKVStore, mod string, data []byte) 
 		}
 		models := make([]weave.Model, len(keys))
 		for i, key := range keys {
-			value, err := db.Get(ix.refKey(key))
+			value, err := db.Get(ix.dbKey(key))
 			if err != nil {
 				return nil, errors.Wrapf(err, "cannot get %q value for %q", i, key)
 			}
@@ -606,8 +614,109 @@ func (ix *nativeIndex) Query(db weave.ReadOnlyKVStore, mod string, data []byte) 
 			}
 		}
 		return models, nil
+	case weave.RangeQueryMod:
+		// Start is the value that was indexed,
+		// Offset is the referenced by this index entity ID,
+		// End is the end indexed value. Often times using end filter
+		// may make no sense, because you cannot know how the index
+		// value is being built.
+		start, offset, end, err := parseIndexQueryRange(data)
+		if err != nil {
+			return nil, errors.Wrap(err, "query data")
+		}
+		if len(end) == 0 {
+			end = bytes.Repeat([]byte{255}, 128) // No limit
+		}
+
+		startKeyChunks := [][]byte{[]byte(ix.name)}
+		if len(offset) > 0 {
+			// If ofset is provided, start must be inserted first,
+			// even if it is nil.
+			startKeyChunks = append(startKeyChunks, start, offset)
+		} else if len(start) > 0 {
+			startKeyChunks = append(startKeyChunks, start)
+		}
+		startKey, err := packNativeIdxKey(startKeyChunks)
+		if err != nil {
+			return nil, errors.Wrap(err, "range start key")
+		}
+		endKey, err := packNativeIdxKey([][]byte{[]byte(ix.name), end})
+		if err != nil {
+			return nil, errors.Wrap(err, "range end key")
+		}
+		it, err := db.Iterator(startKey, endKey)
+		if err != nil {
+			return nil, errors.Wrap(err, "iterator")
+		}
+		return consumeIterator(&paginatedIterator{
+			it: &valueFetchingIterator{
+				db: db,
+				it: &nativeIndexIterator{dbit: it, dbKey: ix.dbKey},
+			},
+			remaining: queryRangeLimit,
+		})
 	default:
 		return nil, errors.Wrap(errors.ErrHuman, "not implemented: "+mod)
+	}
+}
+
+// valueFetchingIterator is an iterator wrapper that fetch value for each
+// returned key. This should be used together with nativeIndexIterator in order
+// to return not only an entity key but also its value.
+type valueFetchingIterator struct {
+	db weave.ReadOnlyKVStore
+	it weave.Iterator
+}
+
+func (v *valueFetchingIterator) Next() (key []byte, value []byte, err error) {
+	key, val, err := v.it.Next()
+	if err != nil {
+		return key, val, err
+	}
+	val, err = v.db.Get(key)
+	return key, val, err
+}
+
+func (v *valueFetchingIterator) Release() {
+	v.it.Release()
+}
+
+// parseIndexQueryRange parse given query data and return range query information.
+// Start and/or end can be nil.
+// Start, end and offset must be hex encoded.
+// Format is <start>[:<offset>[:<end>]] for example:
+//   <start>
+//   <start>:<offset>
+//   <start>:<offset>:
+//   <start>:<offset>:<end>
+//   <start>::<end>
+//   ::<end>
+func parseIndexQueryRange(raw []byte) (start, offset, end []byte, err error) {
+	if len(raw) == 0 {
+		return nil, nil, nil, nil
+	}
+
+	var decErr error // Global decoding error
+	decodeHex := func(b []byte) []byte {
+		if b == nil {
+			return nil
+		}
+		dst := make([]byte, hex.DecodedLen(len(b)))
+		if _, err := hex.Decode(dst, b); err != nil {
+			decErr = errors.Wrap(errors.ErrInput, "not hex data")
+		}
+		return dst
+	}
+
+	switch c := bytes.SplitN(raw, []byte(":"), 4); len(c) {
+	case 1:
+		return decodeHex(raw), nil, nil, decErr
+	case 2:
+		return decodeHex(c[0]), decodeHex(c[1]), nil, decErr
+	case 3:
+		return decodeHex(c[0]), decodeHex(c[1]), decodeHex(c[2]), decErr
+	default:
+		return nil, nil, nil, errors.Wrap(errors.ErrInput, "invalid format")
 	}
 }
 
@@ -615,7 +724,8 @@ func (ix *nativeIndex) Query(db weave.ReadOnlyKVStore, mod string, data []byte) 
 // indexed entities keys. It provides an interface that returns only the
 // relevant data, hiding from the user native index implementation details.
 type nativeIndexIterator struct {
-	dbit weave.Iterator
+	dbit  weave.Iterator
+	dbKey func([]byte) []byte
 }
 
 func (it *nativeIndexIterator) Release() {
@@ -631,10 +741,10 @@ func (it *nativeIndexIterator) Next() ([]byte, []byte, error) {
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "unpack native index key")
 	}
-	return chunks[len(chunks)-1], nil, nil
+	return it.dbKey(chunks[len(chunks)-1]), nil, nil
 }
 
-// packNativeIndexKey serialize a native index key from a set of values to a
+// packNativeIdx serialize a native index key from a set of values to a
 // single key. This process can be reversed using unpackNativeIdxKey function.
 //
 // Native index key is a byte array. After the same for every native index