+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## HEAD
 
+- A new query mod `range` that allows to query for items between certain rage.
+  Result set size is always limited which makes this query suitable for public
+  API.
 - Go 1.13 is now supported and tested by CI
 - Tendermint upgraded to 0.31.11
 - `bnsd/x/termdeposit` a new extension for fix-term deposit of money and